[Drude 1D DNG] - Merged CUDA5.0 into CUDA. - Cutil is completely removed. Current implementation now used CUDA helper libraries and requires CUDA SDK 4.2 or later. - On linux, sm10 architecture is no longer used for compilation. - 'double' is now the default precision.

+++ trunk/Finite.Difference.Time.Domain/Problems/Drude_1D_DNG_Transmission_Coefficient/CUDA/PostProcess.m
@@ -3,7 +3,7 @@
 % Input files are assumed to be in binary format.
 clc
 clear all
-Precision = 'float';
+Precision = 'double';
 
 fidp = fopen ('./FieldData/Parameters.smp', 'r', 'l');
 if fidp == -1
@@ -173,7 +173,7 @@ for i=1:simTime
     if fid == -1
         return;
     end
-    
+
     ExSnapshot = fread (fid, size, Precision);
     figure (6)
     % Scatterer boundaries.
